fix bug and add more s3mi
